refactor(cli): split analyze text rendering into submodule
